# back/central-reserve/shared/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.CreateActionRequest:
    description: Solicitud para crear un nuevo action en el sistema
    properties:
      description:
        example: Permite crear nuevos registros
        type: string
      name:
        example: create
        type: string
    required:
    - name
    type: object
...
  request.UpdateActionRequest:
    description: Solicitud para actualizar un action existente
    properties:
      description:
        example: Permite crear nuevos registros
        type: string
      name:
        example: create
        type: string
    required:
    - name
    type: object
...
  response.LoginErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  response.ResourcePermissions:
    properties:
      actions:
        items:
          type: string
        type: array
      active:
        description: Indica si el recurso está activo para el business
        type: boolean
      resource:
        type: string
    type: object
...
  response.UserRolesPermissionsResponse:
    properties:
      business_id:
        type: integer
      business_name:
        type: string
      business_type_id:
        type: integer
      business_type_name:
        type: string
      is_super:
        type: boolean
      resources:
        items:
          $ref: '#/definitions/response.ResourcePermissions'
        type: array
      role:
        $ref: '#/definitions/services_auth_internal_infra_primary_controllers_authhandler_response.RoleInfo'
    type: object
  response.UserRolesPermissionsSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/response.UserRolesPermissionsResponse'
      success:
        type: boolean
    type: object
...
  services_auth_internal_infra_primary_controllers_authhandler_response.RoleInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:3050
info:
  contact:
    email: backend@example.com
    name: Equipo de Backend
  description: Servicio REST para la gestión de reservas multi-restaurante.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://ejemplo.com/terminos
  title: Restaurant Reservation API
  version: "1.0"
paths:
  /actions:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de actions del sistema con opciones
        de filtrado
      parameters:
      - default: 1
        description: Número de página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Tamaño de página
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Filtrar por nombre (búsqueda parcial)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de actions obtenida exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parámetros de consulta inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener actions
      tags:
      - Actions
    post:
      consumes:
      - application/json
      description: Crea un nuevo action en el sistema con nombre y descripción únicos
      parameters:
      - description: Datos del action a crear
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateActionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Action creado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Solo super usuarios pueden crear actions
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Action ya existe
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear action
      tags:
      - Actions
  /actions/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un action del sistema por su ID único. No se puede eliminar
        si tiene permisos asociados.
      parameters:
      - description: ID del action
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Action eliminado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID de action inválido
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Solo super usuarios pueden eliminar actions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Action no encontrado
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Action tiene permisos asociados
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar action
      tags:
      - Actions
    get:
      consumes:
      - application/json
      description: Obtiene un action específico del sistema por su ID único
      parameters:
      - description: ID del action
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Action obtenido exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID de action inválido
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Action no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener action por ID
      tags:
      - Actions
    put:
      consumes:
      - application/json
      description: Actualiza un action existente en el sistema
      parameters:
      - description: ID del action
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Datos del action a actualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Action actualizado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Solo super usuarios pueden actualizar actions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Action no encontrado
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflicto con action existente
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar action
      tags:
      - Actions
...
  /auth/roles-permissions:
    get:
      consumes:
      - application/json
      description: Obtiene los roles y permisos del usuario autenticado desde el token
      produces:
      - application/json
      responses:
        "200":
          description: Roles y permisos obtenidos exitosamente
          schema:
            $ref: '#/definitions/response.UserRolesPermissionsSuccessResponse'
        "401":
          description: Token de acceso requerido
          schema:
            $ref: '#/definitions/response.LoginErrorResponse'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.LoginErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/response.LoginErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener roles y permisos del usuario
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Token de business JWT con el prefijo **Bearer** (para todos los endpoints)
    in: header
    name: Authorization
    type: apiKey
  BusinessTokenAuth:
    description: Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"